Comment delete added

+++ app/Http/Controllers/FrontEnd/CommentController.php
@@ -50,11 +50,22 @@ class CommentController extends Controller
         if(Auth::check())
         {
             $comment=Comment::where('id',$req->comment_id)->where('user_id',Auth::user()->id)->first();
-            $comment->delete();
+            if($comment)
+            {
+                $comment->delete();
             return response()->json([
                 'status'=>200,
                 'message'=>'Comment Deleted Successfully'
             ]);
+            }
+            else
+            {
+                return response()->json([
+                    'status'=>500,
+                    'message'=>'Something went wrong'
+                ]); 
+            }
+            
         }
         else
         {
